updated decod offset vector as vector of floats

+++ core/api/api.cc
@@ -91,7 +91,9 @@ void pxarCore::setTestboardDelays(std::vector<std::pair<std::string,uint8_t> > s
 }
 
 void pxarCore::setDecodingOffset(uint8_t offset) { _hal->setOffset(offset); }
-void pxarCore::setDecodingOffset(std::vector<uint8_t> offsetv) { _hal->setOffset(offsetv); }
+//void pxarCore::setDecodingOffset(std::vector<uint8_t> offsetv) { _hal->setOffset(offsetv); }
+//void pxarCore::setDecodingOffset(std::vector<uint16_t> offsetv) { _hal->setOffset(offsetv); }
+void pxarCore::setDecodingOffset(std::vector<float> offsetv) { _hal->setOffset(offsetv); }
 
 void pxarCore::setDecodingThresholds(std::vector<std::vector<float> > values) { _hal->setDecodingThresholds(values); }
